+++ R/cloud_local.R
@@ -129,22 +129,12 @@ cloud_object_ls <- function(x, path, extension, prefix = "", suffix = "") {
   check_string(extension)
   check_string(prefix)
   check_string(suffix)
-  
-  if (!grepl("^([A-Za-z]|[0-9]|-|_|\\.|/)+$", path)) {
-    cli::cli_abort(c(
-      "Directory path {.path {path}} is not valid. A valid directory path may \\
-      consist of:",
-      "*" = "uppercase/lowercase letters",
-      "*" = "digits",
-      "*" = "'/' symbols to describe its location inside project's folder",
-      "*" = "'_', '-', '.' symbols or spaces."
-    ))
-  }
+  check_path(path)
   
   if (!grepl("^([A-Za-z]|[0-9])+$", extension)) {
     cli::cli_abort(c(
       "{.arg extension} {.val {extension}} is not valid. A valid extension path \\
-      may consist of:",
+      must consist of:",
       "*" = "uppercase/lowercase letters",
       "*" = "digits"
     ))