+++ R/s3_transfer.R
@@ -3,7 +3,7 @@
 #' @description Uploads a local file from the project's directory to its
 #'   corresponding location within the project's S3 root folder.
 #' 
-#' @inheritParams cloud_validate_file_path
+#' @inheritParams doc_file
 #' @inheritParams cloud_s3_ls
 #'  
 #' @return Invisibly returns `NULL` after successfully uploading the file.
@@ -21,7 +21,7 @@
 #' 
 #' @export
 cloud_s3_upload <- function(file, root = NULL) {
-  cloud_validate_file_path(file)
+  check_path(file)
 
   check_string(root, alt_null = TRUE)
   if (is.null(root)) root <- cloud_s3_get_root()
@@ -35,7 +35,8 @@ cloud_s3_upload <- function(file, root = NULL) {
   aws.s3::put_object(
     bucket = bucket_prefix$bucket,
     file = file,
-    object = bucket_prefix$prefix
+    object = bucket_prefix$prefix,
+    multipart = TRUE
   )
   cli::cli_alert_success(
     "File {.path {file}} uploaded to S3 root {.field {root}}."
@@ -51,7 +52,7 @@ cloud_s3_upload <- function(file, root = NULL) {
 #' @description Retrieves a file from the project's S3 root folder and saves it
 #'   to the local project folder, maintaining the original folder structure.
 #' 
-#' @inheritParams cloud_validate_file_path
+#' @inheritParams doc_file
 #' @inheritParams cloud_s3_ls
 #' 
 #' @return Invisibly returns `NULL` after successfully downloading the file.
@@ -66,7 +67,7 @@ cloud_s3_upload <- function(file, root = NULL) {
 #' 
 #' @export
 cloud_s3_download <- function(file, root = NULL) {
-  cloud_validate_file_path(file)
+  check_path(file)
   
   check_string(root, alt_null = TRUE)
   if (is.null(root)) root <- cloud_s3_get_root()
@@ -90,15 +91,14 @@ cloud_s3_download <- function(file, root = NULL) {
 #'   S3 storage. If no custom writing function is specified, the function will
 #'   infer the appropriate writing method based on the file's extension.
 #'   
-#' @inheritParams cloud_validate_file_path
+#' @inheritParams doc_file
 #' @inheritParams cloud_s3_ls
+#' @inheritParams doc_local
 #' 
 #' @param x An R object to be written to S3.
 #' @param fun A custom writing function. If `NULL` (default), the appropriate
 #'   writing function will be inferred based on the file's extension.
 #' @param ... Additional arguments to pass to the writing function `fun`.
-#' @param local Logical. If `TRUE`, a local copy of the file will also be
-#'   created at the specified path. Default is `FALSE`.
 #' 
 #' @inheritSection cloud_guess_write_fun Default writing functions
 #'
@@ -115,7 +115,7 @@ cloud_s3_download <- function(file, root = NULL) {
 #' @export
 cloud_s3_write <- function(x, file, fun = NULL, ..., local = FALSE,
                          root = NULL) {
-  cloud_validate_file_path(file)
+  check_path(file)
   check_bool(local)
   check_class(fun, "function", alt_null = TRUE)
   check_string(root, alt_null = TRUE)
@@ -141,7 +141,8 @@ cloud_s3_write <- function(x, file, fun = NULL, ..., local = FALSE,
   aws.s3::put_object(
     file = local_file,
     bucket = bucket_prefix$bucket,
-    object = bucket_prefix$prefix
+    object = bucket_prefix$prefix,
+    multipart = TRUE
   )
   
   if (!local) {unlink(local_file)}
@@ -158,7 +159,7 @@ cloud_s3_write <- function(x, file, fun = NULL, ..., local = FALSE,
 #'   function based on the file's extension. However, you can specify a custom
 #'   reading function if necessary.
 #'   
-#' @inheritParams cloud_validate_file_path
+#' @inheritParams doc_file
 #' @inheritParams cloud_s3_ls
 #' 
 #' @param fun A custom reading function. If `NULL` (default), the appropriate
@@ -179,7 +180,7 @@ cloud_s3_write <- function(x, file, fun = NULL, ..., local = FALSE,
 #' 
 #' @export
 cloud_s3_read <- function(file, fun = NULL, ..., root = NULL) {
-  cloud_validate_file_path(file)
+  check_path(file)
   check_string(root, alt_null = TRUE)
   
   if (is.null(fun)) {