+++ NEWS.md
@@ -1,4 +1,9 @@
-# cloudfs (development version)
+# cloudfs 0.1.3
+
+* Updated `cloud_object_ls()` to support paths containing spaces.
+* Improved various sections of the documentation.
+* Enabled multipart upload for S3 writing and uploading functions, with the 
+default part size of 100 MB.
 
 # cloudfs 0.1.2